Add slo tests
