+++ .oxlintrc.json
@@ -11,6 +11,7 @@
 		],
 		"unicorn/require-post-message-target-origin": "off",
 		"yoda": "deny",
+		"vitest/require-mock-type-parameters": "off",
 		"vitest/consistent-test-it": ["error", { "fn": "test" }],
 		"vitest/max-nested-describe": ["error", { "max": 0 }],
 		"vitest/valid-expect": [
