+++ docs/package.json
@@ -12,6 +12,6 @@
 		"start": "vitepress"
 	},
 	"devDependencies": {
-		"vitepress": "^1.6.4"
+		"vitepress": "^2.0.0-alpha.17"
 	}
 }
