+++ package.json
@@ -6,6 +6,7 @@
 	"workspaces": [
 		"docs",
 		"e2e/*",
+		"tests/*",
 		"examples/*",
 		"packages/*",
 		"third-parties/*"
@@ -26,16 +27,16 @@
 		"prepare": "git config core.hooksPath scripts/hooks"
 	},
 	"dependencies": {
-		"@types/node": "^25.3.3"
+		"@types/node": "^25.5.0"
 	},
 	"devDependencies": {
-		"@vitest/coverage-v8": "^4.1.0",
-		"@vitest/ui": "^4.1.0",
-		"oxfmt": "^0.41.0",
-		"oxlint": "^1.56.0",
-		"turbo": "^2.8.12",
+		"@vitest/coverage-v8": "^4.1.2",
+		"@vitest/ui": "^4.1.2",
+		"oxfmt": "^0.43.0",
+		"oxlint": "^1.58.0",
+		"turbo": "^2.9.3",
 		"typescript": "^5.9.3",
-		"vitest": "^4.1.0",
+		"vitest": "^4.1.2",
 		"zx": "^8.8.5"
 	},
 	"overrides": {
