+++ packages/query/package.json
@@ -49,6 +49,7 @@
 	"dependencies": {
 		"@bufbuild/protobuf": "2.10.1",
 		"@ydbjs/api": "^6.0.0",
+		"@ydbjs/abortable": "^6.1.0",
 		"@ydbjs/core": "^6.0.7",
 		"@ydbjs/debug": "^6.0.0",
 		"@ydbjs/error": "^6.0.0",
