+++ tests/slo/Dockerfile
@@ -1,0 +1,10 @@
+# node, oven/bun
+ARG image="node"
+
+FROM ${image}
+WORKDIR /usr/src/app
+
+COPY package.json dist/ ./
+
+ENTRYPOINT [ "node" ]
+CMD [ "kv.test.js" ]
